feat: refactor FilterSelector, SchedulesView, SubjectsView, and GeneratorPage for improved functionality and structure
